Astar is working (sort of), but it's not working well :)

+++ src/ShortCircuit/DataStructures/Interfaces/Heuristic.java
@@ -1,7 +1,7 @@
 package ShortCircuit.DataStructures.Interfaces;
 
 /**
- *
+ * ASKMATTHEW: Generic primitive types
  * @author Development
  */
 public interface Heuristic<T> {
@@ -9,7 +9,7 @@ public interface Heuristic<T> {
     
     public void setEndPosition(T coords);
 
-    public int compareTo(T n);
+    public float compareTo(T n);
